Fixed a couple of bugs in PDPatcher.
Was trying to open the same registrykey twice, and trying to copy files
when they didn't exist when backing up.

+++ PDPatcher/PDPatcher/PDPatcher/FileManager.cs
@@ -20,8 +20,11 @@ namespace PDPatcher
 
             foreach (PatchFile PFile in Manifest.PatchFiles)
             {
-                FileManager.CreateDirectory(WorkingDir + "Backup\\" + PFile.Address);
-                File.Copy(WorkingDir + PFile.Address, WorkingDir + "Backup\\" + PFile.Address);
+                if (File.Exists(WorkingDir + PFile.Address))
+                {
+                    FileManager.CreateDirectory(WorkingDir + "Backup\\" + PFile.Address);
+                    File.Copy(WorkingDir + PFile.Address, WorkingDir + "Backup\\" + PFile.Address);
+                }
             }
         }
